Helper for running queries server-side
