+++ src/index.ts
@@ -1,8 +1,9 @@
-import { postgraphql, createPostGraphQLSchema } from './postgraphql'
+import { postgraphql, createPostGraphQLSchema, withPostGraphQLContext } from './postgraphql'
 
 export default postgraphql
 
 export {
   postgraphql,
   createPostGraphQLSchema,
+  withPostGraphQLContext,
 }
