+++ src/postgraphql/index.ts
@@ -1,7 +1,9 @@
 import postgraphql from './postgraphql'
 import createPostGraphQLSchema from './schema/createPostGraphQLSchema'
+import withPostGraphQLContext from './withPostGraphQLContext'
 
 export {
   postgraphql,
   createPostGraphQLSchema,
+  withPostGraphQLContext,
 }
